bugs with cart fixed.
Users are able to update thier cart information correctly. Cart is now only avaliable in sell page

+++ public/remove_from_cart.php
@@ -13,9 +13,9 @@
     {
         if(isset($_GET["empty"]) && $_GET["empty"] == "true")
         {
-            $sql = "DELETE FROM cart;";
-            $results = query($sql);
-            echo "done";
+            $sql = "DELETE FROM cart WHERE user_id = ?;";
+            $results = query($sql, $_SESSION["id"]);
+            exit;
         } 
     }
